New Version Release
git-svn-id: https://plugins.svn.wordpress.org/gallery-bank/trunk@1114450 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ lib/uninstall-script.php
@@ -1,5 +1,9 @@
 <?php
-global $wpdb;
+global $wpdb,$current_user;
+if (!is_user_logged_in()) {
+	return;
+}
+
 $sql = "DROP TABLE " . gallery_bank_albums();
 $wpdb->query($sql);
